tools/pyproject.toml
[project]
name = "lauschi-catalog"
version = "0.1.0"
description = "Catalog management tools for lauschi (multi-provider Hörspiel catalog)"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "diskcache>=5.6",
    "requests>=2.31",
    "ruamel.yaml>=0.18",
    "rich>=13.7",
    "PyJWT[crypto]>=2.8",
]

[project.optional-dependencies]
ai = [
    "pydantic-ai>=1.62.0",
    "pydantic>=2.0",
]

[project.scripts]
lauschi-catalog = "lauschi_catalog.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lauschi_catalog"]
